pyproject.toml
[project]
name = "project-init"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pexpect>=4.9.0",
    "pyyaml>=6.0.2",
    "questionary>=2.1.0",
    "rich>=14.1.0",
    "typer[all]>=0.16.0",
]

[project.scripts]
# This creates the command-line entry point.
# When installed, typing 'project-init' will run the 'main' function in 'main.py'
project-init = "main:main"